requirements.txt
python-telegram-bot==20.3
pandas==2.0.3
python-dotenv==1.0.0
openpyxl==3.1.2
gunicorn==21.2.0